requirements.txt
aiohttp==3.9.3
discord.py==2.5.0a5122+gaf759857
Flask==2.0.3
pytz==2024.2
Requests==2.32.3
